[feature] Updated to JPA approach. Added lombok. Added custom exceptions

+++ src/main/java/com/finalproject/storeapp/service/ProductsService.java
@@ -8,5 +8,5 @@ public interface ProductsService {
 
     List<Product> findAll();
 
-    Product show(int productId);
+    Product show(long productId);
 }
